refactor: Pawn 클래스와 Board 클래스를 새로운 패키지로 묶음(chess 패키지, pieces 패키지)
