+++ src/main/java/chess/pieces/Pawn.java
@@ -1,9 +1,9 @@
-package chess;
+package chess.pieces;
 
 public class Pawn {
 
-    static final String WHITE_COLOR = "white";
-    static final String BLACK_COLOR = "black";
+    public static final String WHITE_COLOR = "white";
+    public static final String BLACK_COLOR = "black";
 
     private String color;
