+++ src/test/java/chess/BoardTest.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.pieces.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
